HW_5: refactor services: make them more compact
