+++ src/main/java/com/epam/spring/library/service/AuthorService.java
@@ -3,7 +3,6 @@ package com.epam.spring.library.service;
 import com.epam.spring.library.dto.AuthorDTO;
 
 import java.util.List;
-import java.util.Map;
 
 public interface AuthorService {
 
@@ -16,16 +15,4 @@ public interface AuthorService {
     void deleteAuthor(String name);
 
     List<AuthorDTO> getAllAuthors();
-
-    Map<String, String> getAllAuthorNameTranslations(String name);
-
-    Map<String, String> updateAuthorNameTranslation(String name,
-                                                    String langCode,
-                                                    String nameTranslation);
-
-    void clearAuthorNameTranslationsList(String name);
-
-    void deleteAuthorNameTranslationByLangCode(String name, String langCode);
-
-    String getAuthorNameTranslationByLangCode(String name, String langCode);
 }
