+++ src/main/java/com/epam/spring/library/service/BookingService.java
@@ -1,6 +1,5 @@
 package com.epam.spring.library.service;
 
-import com.epam.spring.library.dto.BookDTO;
 import com.epam.spring.library.dto.BookingDTO;
 
 import java.util.List;
@@ -14,14 +13,4 @@ public interface BookingService {
     BookingDTO createBooking(BookingDTO bookingDTO);
 
     BookingDTO updateBooking(int id, BookingDTO bookingDTO);
-
-    List<BookDTO> addBookToBooking(int id, List<BookDTO> bookDTOs);
-
-    void deleteBookFromBooking(int id, String isbn);
-
-    List<BookDTO> getBooksInBooking(int id);
-
-    List<BookDTO> updateBooksListInBooking(int id, List<BookDTO> bookDTOs);
-
-    void clearBookListInBooking(int id);
 }
